ref: Little refactoring
