+++ client/src/components/common/Static/Static.module.scss
@@ -4,6 +4,11 @@
  */
 
 :global(#app) {
+  .button {
+    margin-top: 24px;
+    padding: 0.78571429em 1.5em 0.78571429em;
+  }
+
   .message {
     align-content: space-between;
     align-items: center;
@@ -28,10 +33,6 @@
     line-height: 1.4;
     margin: 4px 0 0;
     text-align: center;
-  }
-
-  .createButton {
-    margin-top: 24px;
   }
 
   .wrapper {
